trying to fix namespace problem with traits
